Steven Kosovich's Web API Homework 1
This is the first assignment, where we downloaded SOAP-UI and got JSON responses for books with "Turing" in the title, and also looked up a country by its IP address. (Note: This currently does not have the results of the assignment, but is just a test by me and will be updated with the results later).