The Postman test results file
Hopefully, this JSON file can be uploaded so you can see the tests and their results that I did with the Google Chrome Postman REST Client. I was confused and I'm not sure if it exported/saved correctly or not. This is the substitute for the SOAPUI tests, since I find Postman a lot easier to use than SOAPUI.